Refactor for some clearer divided responsibilities
